Handle all abilities for Response

+++ app/Http/Controllers/Api/V1/Response/ResponseController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\Api\V1\Response;
 use App\Http\Controllers\Api\V1\Response\FormRequests\StoreResponseRequest;
 use App\Http\Controllers\Api\V1\Response\FormRequests\UpdateResponseRequest;
 use App\Http\Controllers\Controller;
+use App\Models\Account;
+use App\Models\Project;
+use App\Models\Response;
 use App\Services\Api\V1\Responses\ResponseService;
 use Illuminate\Http\JsonResponse;
 
@@ -24,7 +27,20 @@ class ResponseController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Response::class);
         return response()->json($this->responseService->searchResponses());
+    }
+
+    public function ownAccountIndex(int $accountId)
+    {
+        $this->authorize('ownAccountIndex', Account::findOrFail($accountId, ['user_id']));
+        return response()->json($this->responseService->searchByAccount($accountId));
+    }
+
+    public function ownProjectIndex(int $projectId)
+    {
+        $this->authorize('ownProjectIndex', Project::findOrFail($projectId, ['user_id']));
+        return response()->json($this->responseService->searchByProject($projectId));
     }
 
     /**
@@ -35,8 +51,9 @@ class ResponseController extends Controller
      */
     public function store(StoreResponseRequest $request)
     {
+        $this->authorize('create', Response::class);
         $result = $this->responseService->storeResponse($request->getFormData());
-        if (!is_array($result)) {
+        if ($this->isValidationErrorResponse($result)) {
             return response()->json($result, 422);
         }
         return response()->json($result, 201);
@@ -50,6 +67,7 @@ class ResponseController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', $this->responseService->getResponseOnlyAccountAndProject($id));
         return response()->json($this->responseService->findResponse($id));
     }
 
@@ -62,8 +80,9 @@ class ResponseController extends Controller
      */
     public function update(UpdateResponseRequest $request, $id)
     {
+        $this->authorize('update', $this->responseService->getResponseOnlyAccount($id));
         $result = $this->responseService->updateResponse($request->getFormData(), $id);
-        if (!is_array($result)) {
+        if ($this->isValidationErrorResponse($result)) {
             return response()->json($result, 422);
         }
         return response()->json($result);
@@ -77,6 +96,7 @@ class ResponseController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', $this->responseService->getResponseOnlyAccount($id));
         $this->responseService->destroyResponse($id);
         return response()->json([], 204);
     }
